requirements.txt
pandas
numpy
scikit-learn
dill
streamlit
sqlalchemy
mysql-connector-python
-e .